Add basic components
Currently this is just the two column table so we can implement the list of classes as a MCVE.

Note on everything being an interface which we hadn't discussed. When testing higher level components, we need a way to make sure we're not indirectly testing the children. So everything is an interface. :) Dunno if there's a better way to do it.

Change-Id: I245e601a60d847e5eb6e2c4e0b03859d7aa5d2ce