Partially implement docs rendering
There's still a TON of work to do, but this at least gives us the basics.

I didn't add a test for each HTML tag because I felt like that would get out of hand really quickly. I think it's more useful to use the DefaultDocumentationTest as a place to monitor regressions. Anytime something inevitably renders incorrectly, we add a test for that specific edge case. Otherwise we'll just be testing the HTML spec.

Bug: 163811276
Change-Id: I16e871a393832e117916c5bfa11e4b4ee343b035