Refactor link handling
This will allow us to support callable reference links. Currently, we ignore javadoc like [Foo#blah].

BTW, it turns out I was wrong a while back when we were discussing anchor links: Chrome is smart enough to NOT reload the page with an anchor that has a full path. So href="#foo" and href="/path/to/blah#foo" will be treated the same if you're on the blah page.

Change-Id: I78e5d91437db949a7ebab8d4b8ed8b2a5e64c54c