Only compute classlikes once
When I was investigating performance, running `explodedChildren` was the most expensive call in the entire codebase, so I probably over-optimized this. Previously we would call explodedChildren everytime we needed to get classes/interfaces/etc. This CL starts pre-computing those calls at the start of rendering and stores the results for later use.

The idea is that this will be expanded upon in the future to rewrite the documentables or inject new ones. I'm thinking someone would add a `javaClasses` field for example which takes in classList and packages to add the synthetic classes.

Bug: 165105949
Change-Id: Ia20cec702a696a4421d23557e5095b34193fe500