Filter out empty modifiers
JavaModifer.EMPTY appears in the list of modifiers for public methods.
This results in documentation that looks like `public  foo()` with the
extra space after public

Remove any modifier that has an empty name (to cover KotlinModifier.EMPTY as
well. Added tests that check for the number of modifiers in addition to their
presence in a list

BUG: 170742601
Test: ./gradlew
Change-Id: Icf8014b6eee78e94a7fedef00953ea691506584e