Improve text wrapping for lambda parameters
Previously, we were using many &nbsp; characters which made the browser choose arbitrary points for line breaks. By replacing these with regular spaces when appropriate, we give the browser permission to add a line break. In addition, non alphanumeric text - specifically '->', was allowing line breaks in the middle, which is mitigated by wrapping in a span with 'white-space: nowrap;'.

Bug: 180396094
Test: ./gradlew and manual verification of staged changes on devsite
Change-Id: I8639233252e6da4f55ca5e3850a25b10b640515f