Propagate sourceSet around for the sake of `modifiers`.
This will not be used until b/254490320 and we create multiple signatures.
This CL was created as part of the effort to fix all `sourceSet.first()` patterns.
The basis of this CL is because `Projection`s don't have `sourceSet`s,
only `Documentables` do, so I need to pass `sourceSet` as an arg,
because we can't call `CommonOrExpectSourceSet` on a `Projection`.

Test: refactoring-only change;
Change-Id: I302daae7285203004d334b6d019c9770d642c9cf