Display Kotlin properties as accessors in Java docs
The cases I tried to be sure were handled correctly as class/companion/top-level as applicable are regular var, lateinit, @JvmStatic, @JvmField, and const properties, as each interops slightly differently with Java.

This doesn't fully handle top-level/extension properties correctly, as their accessors are not listed as static. These properties are currently not static in the docs, so I made b/268187188 to handle that as a separate issue. This change ensures that properties known to be static will have static accessors.

Test: Updated ClasslikeDocumentableConverterTest tests based on the new behavior, and added some new ones. Updated integration test outputs in a linked CL. Ran against AndroidX tip-of-tree and public to verify there were no errors.
Bug: 261147911

Change-Id: Ie3d345ca0835a9d5ff349f45d380f62a5cabef73