Stop linking to kotlin.coroutines.SuspendFunctionN
`SuspendFunction`s (like `Function`s, which already are not linked to) do not have docs and are not supposed to. They appear only in Java docs and suspend functions are not meant to interop with Java.

This had to be somewhat different from the strategy for not linking to kotlin.jvm.functions.FunctionN because there are valid links for other things in the kotlin.coroutines package.

Bug: 275544850
Test: Added unit test, updated integration tests
Change-Id: I02e9323c90d21c9cf787bb26841c05f3f064a1c6