Add source set data to h2 headers of kmp details sections
This is necessary because the `devsite-filter` around the `<list>` will filter out things not meeting the filter, even if they are `h2`s.

This adds both a `data-title` and a `comment`, following the precedent
of `DefaultPlatformComponent`.

This changes every KMP output file. We may want to delay landing this into its own bugfix release version.

Test: currently staged at
https://android.devsite.corp.google.com/reference/kotlin/androidx/collection/LruCache?db=androidx-refdocs2,
where you can try deselecting `common` to see the H2s vanish.
Fixes: 295937884

Change-Id: I760f013299715897a194029e59faf97e946f3738