Add additional source set fallbacks
aosp/2914706 is failing presubmit because it only has public classes in the desktop and skiko source sets.

This previously used `(this as? DPackage).let`, not `(this as? DPackage)?.let`, which means it was falling back to jvm/android even when the documentable was not a package. Trying to add the `?` led to a failure on androidx-main for a `DTypeAlias` that didn't have a common source set, I assume there are other non-package items that had been using the fallback as well.

Test: locally tested with `./gradlew shadowJar` and cherry-pick of aosp/2914706. `./gradlew docs` ran with no errors
Bug: 254490320
Bug: 322326810
Change-Id: Ibf9672a1757109f480513c4f09c4a136f06a79f7