Refactor isFromSyntheticClass to block better
The old implementation was getting the insert location, running a blocking call, and then adding something.
The new implementation runs the blocking call first

In b/394318482 the only stacktrace involves isFromSyntheticClass; I believe the above fixes what made it uniquely bad.
I believe this fixes the bug, though we'll have to wait and check go/ab-damage-estimator, searching for
`Entry cannot be cast to class java`, in a bit.

Fixes: 394318482
Test: pure-refactor change; no test diff expected or observed. No performance impact.
Change-Id: I90f2e48d6ff99a4cd0a076d49e380a64bd319a69