Drop file-level annotations for DClasslikes
With K1, file-level annotations aren't attached to DClasslikes, but they will be with K2, and the dokka team has said this is undefined behavior.

This change will maintain the current behavior when we switch to K2.

Not including file annotations for classes matters when a file is annotated @RestrictTo: this is meant to apply to the top-level functions and properties in the file, not a class in the file.

Bug: 392589580
Test: ./gradlew test
Change-Id: I86ba2b50d386b8456f73e4413e93af8d5e00c830